Weather_app/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
requests==2.32.3
python-dotenv==1.0.1
SQLAlchemy==2.0.29
pandas==2.2.2
scikit-learn==1.5.1
apscheduler==3.10.4
streamlit==1.36.0
pydantic==2.8.2
matplotlib==3.9.2
